requirements.txt
selenium
webdriver_manager
nltk
python-dotenv
anthropic
